+++ packages/bitcore-wallet-client/lib/api.js
@@ -14,9 +14,6 @@ var Mnemonic = require('bitcore-mnemonic');
 var sjcl = require('sjcl');
 var url = require('url');
 var querystring = require('querystring');
-var Stringify = require('json-stable-stringify');
-
-var request = require('superagent');
 
 var Common = require('./common');
 var Constants = Common.Constants;
@@ -27,8 +24,8 @@ var PayPro = require('./paypro');
 var log = require('./log');
 var Credentials = require('./credentials');
 var Verifier = require('./verifier');
-var Package = require('../package.json');
 var Errors = require('./errors');
+const Request = require('./request');
 
 var BASE_URL = 'http://localhost:3232/bws/api';
 
@@ -41,14 +38,12 @@ var BASE_URL = 'http://localhost:3232/bws/api';
 function API(opts) {
   opts = opts || {};
 
-  this.request = opts.request || request;
-  this.baseUrl = opts.baseUrl || BASE_URL;
-  this.payProHttp = null; // Only for testing
   this.doNotVerifyPayPro = opts.doNotVerifyPayPro;
   this.timeout = opts.timeout || 50000;
   this.logLevel = opts.logLevel || 'silent';
   this.supportStaffWalletId = opts.supportStaffWalletId;
 
+  this.request = new Request(opts.baseUrl || BASE_URL, {r: opts.request});
   log.setLevel(this.logLevel);
 };
 util.inherits(API, events.EventEmitter);
@@ -75,7 +70,7 @@ API.prototype.initialize = function(opts, cb) {
 API.prototype.dispose = function(cb) {
   var self = this;
   self._disposeNotifications();
-  self._logout(cb);
+  self.request.logout(cb);
 };
 
 API.prototype._fetchLatestNotifications = function(interval, cb) {
@@ -217,56 +212,6 @@ API.prototype._processTxps = function(txps) {
 };
 
 /**
- * Parse errors
- * @private
- * @static
- * @memberof Client.API
- * @param {Object} body
- */
-API._parseError = function(body) {
-  if (!body) return;
-
-  if (_.isString(body)) {
-    try {
-      body = JSON.parse(body);
-    } catch (e) {
-      body = {
-        error: body
-      };
-    }
-  }
-  var ret;
-  if (body.code) {
-    if (Errors[body.code]) {
-      ret = new Errors[body.code];
-      if (body.message) ret.message = body.message;
-    } else {
-      ret = new Error(body.code + ': ' + body.message);
-    }
-  } else {
-    ret = new Error(body.error || JSON.stringify(body));
-  }
-  log.error(ret);
-  return ret;
-};
-
-/**
- * Sign an HTTP request
- * @private
- * @static
- * @memberof Client.API
- * @param {String} method - The HTTP method
- * @param {String} url - The URL for the request
- * @param {Object} args - The arguments in case this is a POST/PUT request
- * @param {String} privKey - Private key to sign the request
- */
-API._signRequest = function(method, url, args, privKey) {
-  var message = [method.toLowerCase(), url, JSON.stringify(args)].join('|');
-  return Utils.signMessage(message, privKey);
-};
-
-
-/**
  * Seed from random
  *
  * @param {Object} opts
@@ -279,6 +224,8 @@ API.prototype.seedFromRandom = function(opts) {
 
   opts = opts || {};
   this.credentials = Credentials.create(opts.coin || 'btc', opts.network || 'livenet');
+
+  this.request.setCredentials(this.credentials);
 };
 
 
@@ -369,6 +316,7 @@ API.prototype.seedFromRandomWithMnemonic = function(opts) {
 
   opts = opts || {};
   this.credentials = Credentials.createWithMnemonic(opts.coin || 'btc', opts.network || 'livenet', opts.passphrase, opts.language || 'en', opts.account || 0);
+  this.request.setCredentials(this.credentials);
 };
 
 API.prototype.getMnemonic = function() {
@@ -397,6 +345,7 @@ API.prototype.clearMnemonic = function() {
 API.prototype.seedFromExtendedPrivateKey = function(xPrivKey, opts) {
   opts = opts || {};
   this.credentials = Credentials.fromExtendedPrivateKey(opts.coin || 'btc', xPrivKey, opts.account || 0, opts.derivationStrategy || Constants.DERIVATION_STRATEGIES.BIP44, opts);
+  this.request.setCredentials(this.credentials);
 };
 
 
@@ -417,6 +366,7 @@ API.prototype.seedFromMnemonic = function(words, opts) {
 
   opts = opts || {};
   this.credentials = Credentials.fromMnemonic(opts.coin || 'btc', opts.network || 'livenet', words, opts.passphrase, opts.account || 0, opts.derivationStrategy || Constants.DERIVATION_STRATEGIES.BIP44, opts);
+  this.request.setCredentials(this.credentials);
 };
 
 /**
@@ -435,6 +385,7 @@ API.prototype.seedFromExtendedPublicKey = function(xPubKey, source, entropySourc
 
   opts = opts || {};
   this.credentials = Credentials.fromExtendedPublicKey(opts.coin || 'btc', xPubKey, source, entropySourceHex, opts.account || 0, opts.derivationStrategy || Constants.DERIVATION_STRATEGIES.BIP44);
+  this.request.setCredentials(this.credentials);
 };
 
 
@@ -478,6 +429,7 @@ API.prototype.import = function(str) {
   } catch (ex) {
     throw new Errors.INVALID_BACKUP;
   }
+  this.request.setCredentials(this.credentials);
 };
 
 API.prototype._import = function(cb) {
@@ -542,6 +494,7 @@ API.prototype.importFromMnemonic = function(words, opts, cb) {
     log.info('Mnemonic error:', e);
     return cb(new Errors.INVALID_BACKUP);
   }
+  this.request.setCredentials(this.credentials);
 
   self._import(function(err, ret) {
     if (!err) return cb(null, ret);
@@ -550,6 +503,7 @@ API.prototype.importFromMnemonic = function(words, opts, cb) {
       var altCredentials = derive(true);
       if (altCredentials.xPubKey.toString() == self.credentials.xPubKey.toString()) return cb(err);
       self.credentials = altCredentials;
+      self.request.setCredentials(self.credentials);
       return self._import(cb);
     }
     return cb(err);
@@ -583,6 +537,7 @@ API.prototype.importFromExtendedPrivateKey = function(xPrivKey, opts, cb) {
     return cb(new Errors.INVALID_BACKUP);
   };
 
+  this.request.setCredentials(this.credentials);
   this._import(cb);
 };
 
@@ -612,6 +567,7 @@ API.prototype.importFromExtendedPublicKey = function(xPubKey, source, entropySou
     return cb(new Errors.INVALID_BACKUP);
   };
 
+  this.request.setCredentials(this.credentials);
   this._import(cb);
 };
 
@@ -717,7 +673,7 @@ API.prototype.openWallet = function(cb) {
   if (self.credentials.isComplete() && self.credentials.hasWalletInfo())
     return cb(null, true);
 
-  self._doGetRequest('/v2/wallets/?includeExtendedInfo=1', function(err, ret) {
+  self.request.get('/v2/wallets/?includeExtendedInfo=1', function(err, ret) {
     if (err) return cb(err);
     var wallet = ret.wallet;
 
@@ -750,194 +706,6 @@ API.prototype.openWallet = function(cb) {
   });
 };
 
-API.prototype._getHeaders = function(method, url, args) {
-  var headers = {
-    'x-client-version': 'bwc-' + Package.version,
-  };
-  if (this.supportStaffWalletId) {
-    headers['x-wallet-id'] = this.supportStaffWalletId;
-  }
-
-  return headers;
-};
-
-
-/**
- * Do an HTTP request
- * @private
- *
- * @param {Object} method
- * @param {String} url
- * @param {Object} args
- * @param {Callback} cb
- */
-API.prototype._doRequest = function(method, url, args, useSession, cb) {
-  var self = this;
-
-  var headers = self._getHeaders(method, url, args);
-
-  if (self.credentials) {
-    headers['x-identity'] = self.credentials.copayerId;
-
-    if (useSession && self.session) {
-      headers['x-session'] = self.session;
-    } else {
-      var reqSignature;
-      var key = args._requestPrivKey || self.credentials.requestPrivKey;
-      if (key) {
-        delete args['_requestPrivKey'];
-        reqSignature = API._signRequest(method, url, args, key);
-      }
-      headers['x-signature'] = reqSignature;
-    }
-  }
-
-  var r = self.request[method](self.baseUrl + url);
-
-  r.accept('json');
-
-  _.each(headers, function(v, k) {
-    if (v) r.set(k, v);
-  });
-
-  if (args) {
-    if (method == 'post' || method == 'put') {
-      r.send(args);
-
-    } else {
-      r.query(args);
-    }
-  }
-
-  r.timeout(self.timeout);
-
-  r.end(function(err, res) {
-    if (!res) {
-      return cb(new Errors.CONNECTION_ERROR);
-    }
-
-    if (res.body)
-
-      log.debug(util.inspect(res.body, {
-        depth: 10
-      }));
-
-    if (res.status !== 200) {
-      if (res.status === 404)
-        return cb(new Errors.NOT_FOUND);
-
-      if (!res.status)
-        return cb(new Errors.CONNECTION_ERROR);
-
-      log.error('HTTP Error:' + res.status);
-
-      if (!res.body)
-        return cb(new Error(res.status));
-
-      return cb(API._parseError(res.body));
-    }
-
-    if (res.body === '{"error":"read ECONNRESET"}')
-      return cb(new Errors.ECONNRESET_ERROR(JSON.parse(res.body)));
-
-    return cb(null, res.body, res.header);
-  });
-};
-
-API.prototype._login = function(cb) {
-  this._doPostRequest('/v1/login', {}, cb);
-};
-
-API.prototype._logout = function(cb) {
-  this._doPostRequest('/v1/logout', {}, cb);
-};
-
-/**
- * Do an HTTP request
- * @private
- *
- * @param {Object} method
- * @param {String} url
- * @param {Object} args
- * @param {Callback} cb
- */
-API.prototype._doRequestWithLogin = function(method, url, args, cb) {
-  var self = this;
-
-  function doLogin(cb) {
-    self._login(function(err, s) {
-      if (err) return cb(err);
-      if (!s) return cb(new Errors.NOT_AUTHORIZED);
-      self.session = s;
-      cb();
-    });
-  };
-
-  async.waterfall([
-
-    function(next) {
-      if (self.session) return next();
-      doLogin(next);
-    },
-    function(next) {
-      self._doRequest(method, url, args, true, function(err, body, header) {
-        if (err && err instanceof Errors.NOT_AUTHORIZED) {
-          doLogin(function(err) {
-            if (err) return next(err);
-            return self._doRequest(method, url, args, true, next);
-          });
-        }
-        next(null, body, header);
-      });
-    },
-  ], cb);
-};
-
-/**
- * Do a POST request
- * @private
- *
- * @param {String} url
- * @param {Object} args
- * @param {Callback} cb
- */
-API.prototype._doPostRequest = function(url, args, cb) {
-  return this._doRequest('post', url, args, false, cb);
-};
-
-API.prototype._doPutRequest = function(url, args, cb) {
-  return this._doRequest('put', url, args, false, cb);
-};
-
-/**
- * Do a GET request
- * @private
- *
- * @param {String} url
- * @param {Callback} cb
- */
-API.prototype._doGetRequest = function(url, cb) {
-  url += url.indexOf('?') > 0 ? '&' : '?';
-  url += 'r=' + _.random(10000, 99999);
-  return this._doRequest('get', url, {}, false, cb);
-};
-
-API.prototype._doGetRequestWithLogin = function(url, cb) {
-  url += url.indexOf('?') > 0 ? '&' : '?';
-  url += 'r=' + _.random(10000, 99999);
-  return this._doRequestWithLogin('get', url, {}, cb);
-};
-
-/**
- * Do a DELETE request
- * @private
- *
- * @param {String} url
- * @param {Callback} cb
- */
-API.prototype._doDeleteRequest = function(url, cb) {
-  return this._doRequest('delete', url, {}, false, cb);
-};
 
 API._buildSecret = function(walletId, walletPrivKey, coin, network) {
   if (_.isString(walletPrivKey)) {
@@ -1122,7 +890,7 @@ API.prototype._doJoinWallet = function(walletId, walletPrivKey, xPubKey, request
   args.copayerSignature = Utils.signMessage(hash, walletPrivKey);
 
   var url = '/v2/wallets/' + walletId + '/copayers';
-  this._doPostRequest(url, args, function(err, body) {
+  this.request.post(url, args, function(err, body) {
     if (err) return cb(err);
     self._processWallet(body.wallet);
     return cb(null, body.wallet);
@@ -1244,7 +1012,7 @@ API.prototype.getFeeLevels = function(coin, network, cb) {
   $.checkArgument(coin || _.includes(['btc', 'bch'], coin));
   $.checkArgument(network || _.includes(['livenet', 'testnet'], network));
 
-  self._doGetRequest('/v2/feelevels/?coin=' + (coin || 'btc') + '&network=' + (network || 'livenet'), function(err, result) {
+  self.request.get('/v2/feelevels/?coin=' + (coin || 'btc') + '&network=' + (network || 'livenet'), function(err, result) {
     if (err) return cb(err);
     return cb(err, result);
   });
@@ -1256,7 +1024,7 @@ API.prototype.getFeeLevels = function(coin, network, cb) {
  * @param {Callback} cb
  */
 API.prototype.getVersion = function(cb) {
-  this._doGetRequest('/v1/version/', cb);
+  this.request.get('/v1/version/', cb);
 };
 
 API.prototype._checkKeyDerivation = function() {
@@ -1331,7 +1099,7 @@ API.prototype.createWallet = function(walletName, copayerName, m, n, opts, cb) {
     singleAddress: !!opts.singleAddress,
     id: opts.id,
   };
-  self._doPostRequest('/v2/wallets/', args, function(err, res) {
+  self.request.post('/v2/wallets/', args, function(err, res) {
     if (err) return cb(err);
 
     var walletId = res.walletId;
@@ -1441,7 +1209,7 @@ API.prototype.recreateWallet = function(cb) {
       supportBIP44AndP2PKH: supportBIP44AndP2PKH,
     };
 
-    self._doPostRequest('/v2/wallets/', args, function(err, body) {
+    self.request.post('/v2/wallets/', args, function(err, body) {
       if (err) {
         if (!(err instanceof Errors.WALLET_ALREADY_EXISTS))
           return cb(err);
@@ -1558,7 +1326,7 @@ API.prototype.getNotifications = function(opts, cb) {
     url += '?timeSpan=' + opts.timeSpan;
   }
 
-  self._doGetRequestWithLogin(url, function(err, result) {
+  self.request.getWithLogin(url, function(err, result) {
     if (err) return cb(err);
 
     var notifications = _.filter(result, function(notification) {
@@ -1592,7 +1360,7 @@ API.prototype.getStatus = function(opts, cb) {
   qs.push('includeExtendedInfo=' + (opts.includeExtendedInfo ? '1' : '0'));
   qs.push('twoStep=' + (opts.twoStep ? '1' : '0'));
 
-  self._doGetRequest('/v2/wallets/?' + qs.join('&'), function(err, result) {
+  self.request.get('/v2/wallets/?' + qs.join('&'), function(err, result) {
     if (err) return cb(err);
     if (result.wallet.status == 'pending') {
       var c = self.credentials;
@@ -1616,7 +1384,7 @@ API.prototype.getPreferences = function(cb) {
   $.checkArgument(cb);
 
   var self = this;
-  self._doGetRequest('/v1/preferences/', function(err, preferences) {
+  self.request.get('/v1/preferences/', function(err, preferences) {
     if (err) return cb(err);
     return cb(null, preferences);
   });
@@ -1634,7 +1402,7 @@ API.prototype.savePreferences = function(preferences, cb) {
   $.checkArgument(cb);
 
   var self = this;
-  self._doPutRequest('/v1/preferences/', preferences, cb);
+  self.request.put('/v1/preferences/', preferences, cb);
 };
 
 
@@ -1654,8 +1422,11 @@ API.prototype.fetchPayPro = function(opts, cb) {
  
   PayPro.get({
     url: opts.payProUrl,
-    http: this.payProHttp,
     coin: this.credentials.coin || 'btc',
+    network: this.credentials.network || 'livenet',
+
+    // for testing
+    request: this.request,
   }, function(err, paypro) { 
     if (err)
       return cb(err);
@@ -1681,7 +1452,7 @@ API.prototype.getUtxos = function(opts, cb) {
       addresses: [].concat(opts.addresses).join(',')
     });
   }
-  this._doGetRequest(url, cb);
+  this.request.get(url, cb);
 };
 
 API.prototype._getCreateTxProposalArgs = function(opts) {
@@ -1729,7 +1500,7 @@ API.prototype.createTxProposal = function(opts, cb) {
 
   var args = self._getCreateTxProposalArgs(opts);
 
-  self._doPostRequest('/v2/txproposals/', args, function(err, txp) {
+  self.request.post('/v2/txproposals/', args, function(err, txp) {
     if (err) return cb(err);
 
     self._processTxps(txp);
@@ -1764,7 +1535,7 @@ API.prototype.publishTxProposal = function(opts, cb) {
   };
 
   var url = '/v1/txproposals/' + opts.txp.id + '/publish/';
-  self._doPostRequest(url, args, function(err, txp) {
+  self.request.post(url, args, function(err, txp) {
     if (err) return cb(err);
     self._processTxps(txp);
     return cb(null, txp);
@@ -1794,7 +1565,7 @@ API.prototype.createAddress = function(opts, cb) {
 
   opts = opts || {};
 
-  self._doPostRequest('/v3/addresses/', opts, function(err, address) {
+  self.request.post('/v3/addresses/', opts, function(err, address) {
     if (err) return cb(err);
 
     if (!Verifier.checkAddress(self.credentials, address)) {
@@ -1831,7 +1602,7 @@ API.prototype.getMainAddresses = function(opts, cb) {
   }
   var url = '/v1/addresses/' + qs;
 
-  self._doGetRequest(url, function(err, addresses) {
+  self.request.get(url, function(err, addresses) {
     if (err) return cb(err);
 
     if (!opts.doNotVerify) {
@@ -1849,7 +1620,6 @@ API.prototype.getMainAddresses = function(opts, cb) {
  * Update wallet balance
  *
  * @param {String} opts.coin - Optional: defaults to current wallet coin
- * @param {Boolean} opts.twoStep[=false] - Optional: use 2-step balance computation for improved performance
  * @param {Callback} cb
  */
 API.prototype.getBalance = function(opts, cb) {
@@ -1865,7 +1635,6 @@ API.prototype.getBalance = function(opts, cb) {
   $.checkState(this.credentials && this.credentials.isComplete());
 
   var args = [];
-  if (opts.twoStep) args.push('?twoStep=1');
   if (opts.coin) {
     if (!_.includes(['btc', 'bch'], opts.coin)) return cb(new Error('Invalid coin'));
     args.push('coin=' + opts.coin);
@@ -1876,7 +1645,7 @@ API.prototype.getBalance = function(opts, cb) {
   }
 
   var url = '/v1/balance/' + qs;
-  this._doGetRequest(url, cb);
+  this.request.get(url, cb);
 };
 
 /**
@@ -1893,7 +1662,7 @@ API.prototype.getTxProposals = function(opts, cb) {
 
   var self = this;
 
-  self._doGetRequest('/v1/txproposals/', function(err, txps) {
+  self.request.get('/v1/txproposals/', function(err, txps) {
     if (err) return cb(err);
 
     self._processTxps(txps);
@@ -1901,7 +1670,6 @@ API.prototype.getTxProposals = function(opts, cb) {
       function(txp, acb) {
         if (opts.doNotVerify) return acb(true);
         self.getPayPro(txp, function(err, paypro) {
-
           var isLegit = Verifier.checkTxProposal(self.credentials, txp, {
             paypro: paypro,
           });
@@ -1939,10 +1707,13 @@ API.prototype.getPayPro = function(txp, cb) {
 
   PayPro.get({
     url: txp.payProUrl,
-    http: self.payProHttp,
     coin: txp.coin || 'btc',
+    network: txp.network || 'livenet',
+
+    // for testing
+    request: self.request,
   }, function(err, paypro) {
-    if (err) return cb(new Error('Cannot check transaction now:' + err));
+    if (err) return cb(new Error('Could not fetch invoice:' + (err.message? err.message : err)));
     return cb(null, paypro);
   });
 };
@@ -2001,7 +1772,7 @@ API.prototype.signTxProposal = function(txp, password, cb) {
       signatures: signatures
     };
 
-    self._doPostRequest(url, args, function(err, txp) {
+    self.request.post(url, args, function(err, txp) {
       if (err) return cb(err);
       self._processTxps(txp);
       return cb(null, txp);
@@ -2132,7 +1903,7 @@ API.prototype.rejectTxProposal = function(txp, reason, cb) {
   var args = {
     reason: API._encryptMessage(reason, self.credentials.sharedEncryptingKey) || '',
   };
-  self._doPostRequest(url, args, function(err, txp) {
+  self.request.post(url, args, function(err, txp) {
     if (err) return cb(err);
     self._processTxps(txp);
     return cb(null, txp);
@@ -2157,7 +1928,7 @@ API.prototype.broadcastRawTx = function(opts, cb) {
   opts = opts || {};
 
   var url = '/v1/broadcast_raw/';
-  self._doPostRequest(url, opts, function(err, txid) {
+  self.request.post(url, opts, function(err, txid) {
     if (err) return cb(err);
     return cb(null, txid);
   });
@@ -2166,7 +1937,7 @@ API.prototype.broadcastRawTx = function(opts, cb) {
 API.prototype._doBroadcast = function(txp, cb) {
   var self = this;
   var url = '/v1/txproposals/' + txp.id + '/broadcast/';
-  self._doPostRequest(url, {}, function(err, txp) {
+  self.request.post(url, {}, function(err, txp) {
     if (err) return cb(err);
     self._processTxps(txp);
     return cb(null, txp);
@@ -2187,26 +1958,33 @@ API.prototype.broadcastTxProposal = function(txp, cb) {
   var self = this;
 
   self.getPayPro(txp, function(err, paypro) {
+    if (err) return cb(err);
 
     if (paypro) {
 
+      var t_unsigned = Utils.buildTx(txp);
       var t = Utils.buildTx(txp);
       self._applyAllSignatures(txp, t);
 
       PayPro.send({
-        http: self.payProHttp,
         url: txp.payProUrl,
         amountSat: txp.amount,
-        refundAddr: txp.changeAddress.address,
-        merchant_data: paypro.merchant_data,
+        rawTxUnsigned: t_unsigned.uncheckedSerialize(),
         rawTx: t.serialize({
           disableSmallFees: true,
           disableLargeFees: true,
           disableDustOutputs: true
         }),
         coin: txp.coin || 'btc',
+        network: txp.network || 'livenet',
+
+        // for testing
+        request: self.request,
       }, function(err, ack, memo) {
-        log.warn('Merchant rejected the payment. Broadcasting it any ways.', err);
+        if (err) {
+          return cb(err);
+        }
+
         if (memo) {
           log.debug('Merchant memo:', memo);
         }
@@ -2233,7 +2011,7 @@ API.prototype.removeTxProposal = function(txp, cb) {
   var self = this;
 
   var url = '/v1/txproposals/' + txp.id;
-  self._doDeleteRequest(url, function(err) {
+  self.request.delete(url, function(err) {
     return cb(err);
   });
 };
@@ -2264,7 +2042,7 @@ API.prototype.getTxHistory = function(opts, cb) {
   }
 
   var url = '/v1/txhistory/' + qs;
-  self._doGetRequest(url, function(err, txs) {
+  self.request.get(url, function(err, txs) {
     if (err) return cb(err);
     self._processTxps(txs);
     return cb(null, txs);
@@ -2282,7 +2060,7 @@ API.prototype.getTx = function(id, cb) {
 
   var self = this;
   var url = '/v1/txproposals/' + id;
-  this._doGetRequest(url, function(err, txp) {
+  this.request.get(url, function(err, txp) {
     if (err) return cb(err);
 
     self._processTxps(txp);
@@ -2308,7 +2086,7 @@ API.prototype.startScan = function(opts, cb) {
     includeCopayerBranches: opts.includeCopayerBranches,
   };
 
-  self._doPostRequest('/v1/addresses/scan', args, function(err) {
+  self.request.post('/v1/addresses/scan', args, function(err) {
     return cb(err);
   });
 };
@@ -2347,7 +2125,7 @@ API.prototype.addAccess = function(opts, cb) {
     restrictions: opts.restrictions,
   };
 
-  this._doPutRequest('/v1/copayers/' + copayerId + '/', opts, function(err, res) {
+  this.request.put('/v1/copayers/' + copayerId + '/', opts, function(err, res) {
     if (err) return cb(err);
     return cb(null, res.wallet, reqPrivKey);
   });
@@ -2364,7 +2142,7 @@ API.prototype.getTxNote = function(opts, cb) {
   var self = this;
 
   opts = opts || {};
-  self._doGetRequest('/v1/txnotes/' + opts.txid + '/', function(err, note) {
+  self.request.get('/v1/txnotes/' + opts.txid + '/', function(err, note) {
     if (err) return cb(err);
     self._processTxNotes(note);
     return cb(null, note);
@@ -2386,7 +2164,7 @@ API.prototype.editTxNote = function(opts, cb) {
   if (opts.body) {
     opts.body = API._encryptMessage(opts.body, this.credentials.sharedEncryptingKey);
   }
-  self._doPutRequest('/v1/txnotes/' + opts.txid + '/', opts, function(err, note) {
+  self.request.put('/v1/txnotes/' + opts.txid + '/', opts, function(err, note) {
     if (err) return cb(err);
     self._processTxNotes(note);
     return cb(null, note);
@@ -2413,7 +2191,7 @@ API.prototype.getTxNotes = function(opts, cb) {
     qs = '?' + args.join('&');
   }
 
-  self._doGetRequest('/v1/txnotes/' + qs, function(err, notes) {
+  self.request.get('/v1/txnotes/' + qs, function(err, notes) {
     if (err) return cb(err);
     self._processTxNotes(notes);
     return cb(null, notes);
@@ -2443,7 +2221,7 @@ API.prototype.getFiatRate = function(opts, cb) {
     qs = '?' + args.join('&');
   }
 
-  self._doGetRequest('/v1/fiatrates/' + opts.code + '/' + qs, function(err, rates) {
+  self.request.get('/v1/fiatrates/' + opts.code + '/' + qs, function(err, rates) {
     if (err) return cb(err);
     return cb(null, rates);
   });
@@ -2458,7 +2236,7 @@ API.prototype.getFiatRate = function(opts, cb) {
  */
 API.prototype.pushNotificationsSubscribe = function(opts, cb) {
   var url = '/v1/pushnotifications/subscriptions/';
-  this._doPostRequest(url, opts, function(err, response) {
+  this.request.post(url, opts, function(err, response) {
     if (err) return cb(err);
     return cb(null, response);
   });
@@ -2471,7 +2249,7 @@ API.prototype.pushNotificationsSubscribe = function(opts, cb) {
  */
 API.prototype.pushNotificationsUnsubscribe = function(token, cb) {
   var url = '/v2/pushnotifications/subscriptions/' + token;
-  this._doDeleteRequest(url, cb);
+  this.request.delete(url, cb);
 };
 
 /**
@@ -2482,7 +2260,7 @@ API.prototype.pushNotificationsUnsubscribe = function(token, cb) {
  */
 API.prototype.txConfirmationSubscribe = function(opts, cb) {
   var url = '/v1/txconfirmations/';
-  this._doPostRequest(url, opts, function(err, response) {
+  this.request.post(url, opts, function(err, response) {
     if (err) return cb(err);
     return cb(null, response);
   });
@@ -2495,7 +2273,7 @@ API.prototype.txConfirmationSubscribe = function(opts, cb) {
  */
 API.prototype.txConfirmationUnsubscribe = function(txid, cb) {
   var url = '/v1/txconfirmations/' + txid;
-  this._doDeleteRequest(url, cb);
+  this.request.delete(url, cb);
 };
 
 /**
@@ -2524,7 +2302,7 @@ API.prototype.getSendMaxInfo = function(opts, cb) {
 
   var url = '/v1/sendmaxinfo/' + qs;
 
-  self._doGetRequest(url, function(err, result) {
+  self.request.get(url, function(err, result) {
     if (err) return cb(err);
     return cb(null, result);
   });
@@ -2548,7 +2326,7 @@ API.prototype.getStatusByIdentifier = function(opts, cb) {
   qs.push('includeExtendedInfo=' + (opts.includeExtendedInfo ? '1' : '0'));
   qs.push('walletCheck=' + (opts.walletCheck ? '1' : '0'));
 
-  self._doGetRequest('/v1/wallets/' + opts.identifier + '?' + qs.join('&'), function(err, result) {
+  self.request.get('/v1/wallets/' + opts.identifier + '?' + qs.join('&'), function(err, result) {
     if (err || !result || !result.wallet) return cb(err);
     if (result.wallet.status == 'pending') {
       var c = self.credentials;
@@ -2622,7 +2400,10 @@ API.prototype.createWalletFromOldCopay = function(username, password, blob, cb) 
     return cb(new Error('Wallet is incomplete, cannot be imported'));
 
   this.credentials = Credentials.fromOldCopayWallet(w);
+  this.request.setCredentials(this.credentials);
   this.recreateWallet(cb);
 };
+
+API.PayPro = PayPro;
 
 module.exports = API;