+++ packages/bitcore-wallet-service/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.1.1](https://github.com/bitpay/bitcore-wallet-service/compare/v8.1.0...v8.1.1) (2019-03-21)
+
+**Note:** Version bump only for package bitcore-wallet-service
+
+
+
+
+
 # [8.1.0](https://github.com/bitpay/bitcore-wallet-service/compare/v5.0.0-beta.44...v8.1.0) (2019-02-27)